Fixup card alignment in multiple places.
Move the left margin required for the drawer out of the paged list
view layout since this was causing other lists (for example the ones
using StreamCardView) to not be centered.

The app:rightGutterEnabled attribute is also fairly useless and seems
to do nothing corresponding to its description. Instead, remove that
attribute altogether and replace it with app:usedInDrawer which then
sets the margin (which we removed above) in the code. This way, lists
that live in the drawer can look correct while lists that don't live
in a drawer can also look right.

Test: Checked drawer and card stream alignment in dialer, media
      settings & lenspicker.
Change-Id: Id3a333ac798984841f9506375792b0358cddc60a