Fix reference to package in AndroidManifest.
The package references stream.ui. This directory has been deleted.
Instead, use a package that exists.

Ideally, the package would be a common root directory for all files in
this directory, but it is a very generic com.android.car. It would be
best not to use that as the package name.

This directory will eventually be deleted anyway when the car support
library is finished.

Also, rename some layouts to avoid resource overriding.

Test: build radio, dialer and LensPicker to ensure no errors.
Bug: 67116311
Change-Id: I098fe78bb475f1d19835bdabaadb172a2ace0654