DO NOT MERGE Clean deps of car-apps-common and car-stream-ui-lib.
This refactor removes all support library dependencies from
car-apps-common. This allows dependent applications to migrate to the
new androidx namespace more easily.

In addition, remove the number of shared components in car-stream-ui-lib
to just the background resolver. This reduces the scope of this library
and the number of dependent applications. Also, needed to update the
AndroidManifest to have the correct package name.

Lastly, move ColorChecker to car-media-common because that's where it's
being used. This also allows the support library dep to be removed
from car-apps-common.

Bug: 78886311
Test: build dependent applications and deploy.
Change-Id: I7e287c819c4fe36524f10345b58b67bfbd65fced