Fix ManifestMerger issues
Set LOCAL_MIN_SDK_VERSION in car-settings-lib so that the build system
doesn't raise the minSdkVersion value specified in the manifest, which
causes ManifestMerger to error when trying to merge it into a package
that specifies a lower minSdkVersion.

Remove the empty application with an android:label from
car-stream-ui-lib which causes ManifestMerger to error when combining
two applications with different labels.

Bug: 78447299
Test: m java
Change-Id: I12230134d53a949a0e52dba7224cbec882fd8380