Move ActionBar from the support lib to Car package
This CL adds new files to Car/libs and renames ActionBar to
CarActionBar.

Also it adds 2 new changes in functionality:
1. Added new attribute "enableOverflow" to control availability of the
overflow button (it will prevent overflow button to appear on small
layouts when back/prev buttons are hidden)
2. It uses special extras for preserving Prev/Next controls and can hide
those if not reserved (to reuse the space for custom actions).

b/117368437

Test: manual testing



Change-Id: Iff74379733d40a5ad0036f77bdf1095c8c207f39