Don't override getLayoutManager()
Instead provide a new getEffectiveLayoutManager() to provide the inner
recycler view's layout manager.

We can't override getLayoutManager directly because it affects the behavior of
the accessibility stack. In the long term, PagedRecyclerView should be
modified to handle all of these corner cases to act as a drop in for
RecyclerViews.

Bug: 135625490
Test: manual, make AndroidAutoUiTests + SettingHelperTest
Change-Id: I19112cfe2e36d7e06f1f84d1c15d9a27ebad3ab8