Prevent scrolling in outer LM when scrollbar shown
For PagedRecyclerView, when the scorllbar is shown, the nested
RecyclerView needs to receive touch events for moving SeekBars in child
views. However, because the outer LayoutManager supported scrolling, it
would intercept touch events once it received a sufficient move event.
This made it so seekbars would not receive move events rendering them
unresponsive.

Fixes: 135036948

Test: build and deploy, move seekbars within scroll bar enabled PRV
Change-Id: Id7bf7a85e7160ee2706b75166a8c351195d9c93c