Refactor image handling
Note: still a few rough edges (terse javadoc and a few "ABNOWs" I want
to address before submission.

The code of this CL is LocalImageFetcher that fetches images with the
new framework's ImageDecoder (meant to replace glide), coalescing multiple
requests for the same image into a single task, and caching the results in
memory.

ImageBinder and ImageViewBinder are the helper objects to use to bind
images to UI elements asynchronously.


Test: manual
Fixes: 134943941
Change-Id: Ic834ac0edd6905ec36388ca9657f07a4321af4e4