Center scroll bar for the reference design.
Scroll bar thumb is always centered withing the width of the scroll bar view. View.layout() always takes the integer value and before this change the width of thumb was even and the width of scroll bar view was odd. Hence we were seeing that scroll bar was 1 pixel off. Now making both the values odd so it can be properly centered.

Bug: 134156303
Test: Manual
Change-Id: Ie781dd9349200067cafe83ba884a57a689ac64c7