Fix search box sometimes not losing focus
Our SearchView calls clearFocus() on the EditText when the search/done
button on the keyboard is pressed. clearFocus() doesn't actually clear
focus, it just selects the first focusable item in the view hiearchy.
This could potentially be the same EditText, causing focus to not be lost.

This also causes a SysUi crash in the current implementation of SysUi.
Hiding and then immediately reshowing the keyboard causes a crash
related to the functionality to hide the navbar while the keyboard is open.

Bug: 141478974
Test: Manually
Change-Id: Ic0ac761b2a1a93cb699326a7c0fd6a83692cfe53