Set setHasFixedSize property to true for nested RV
According to the java doc:
"RecyclerView can perform several optimizations if it can know in
advance that RecyclerView's size is not affected by the adapter
contents. RecyclerView can still change its size based on other
factors (e.g. its parent's size) but this size calculation cannot
depend on the size of its children or contents of its adapter
(except the number of items in the adapter). If your use of RecyclerView
falls into this category, set this to {@code true}. It will allow
RecyclerView to avoid invalidating the whole layout when its adapter
contents change."

Our use case lie within this as the nested recycler view will always
take the height of the parent RV (CarUiRecyclerView) defined by
the application using it.

Bug: 136669451
Test: Manual, robolectric
Change-Id: I8504281cfa2756885cc30e3fb8afc4c7ebe9b376