Prevent one app from blocking all download threads + fix getIconResource
Since we can't cancel tasks without requesting developers to handle it,
this CL groups loading tasks for the same application into the same thread
pool, and keeps a cache of pools with up to
image_fetcher_thread_pools_max_count entries.

Fix UriUtils.getIconResource that was incorrectly equating the authority
of a content provider uri with the package name.

Fixes: 140265697
Test: manual
Change-Id: I8378ff0efdb06024c04c11187cecc61b47007049