DO NOT MERGE: Add subtitles to Chassis' AlertDialogBuilder
This involves setting a custom view as a title. This custom view has to
mimic our system's dialog titles as closely as possible, so I brought in
some keylines and stuff from our car_product dialog overlay.

This is DO NOT MERGE because of a merge conflict in master. Will resolve
and synchronize the branches again as part of the chain of cherry-picks.

Fixes: 145164596
Test: Manually
Change-Id: I88d77968a70019af8c22abd9936597c92153462c