Fix memory leak when retaining MenuItems
If a MenuItem is added to a toolbar, and then we leave the screen with
the toolbar but hold on to the MenuItem, the screen with the toolbar
won't be garbage-collected because the MenuItem holds references to it.

Even after this chagne, it's still very easy to leak activities
if your MenuItem's onClickListener uses the activity. Clients
must be careful of that.

Test: In Paintbooth's ToolbarActivity, add a static list of MenuItems and
      add the settings MenuItem to the list when it's created. Also change
      the settings OnClickListener to use the application context for the
      toast, calling getApplicationContext() outside the lamda. Then going
      to the ToolbarActivity and back to the MainActivity 5 times
      will cause LeakCanary to trigger without this change. Adding the
      search item will cause a leak even after this change because its
      OnClickListener references the toolbar.
Change-Id: Ie923d92f8363427aacb177a52690519aaf9f5433