Fix issues with association flow
Found when enrolling a device previously unseen:

Name retrieval is asynchronous, therefore the entity must
support a null name and write once received.

Callbacks were not issued properly so features were not
made aware that a new device had connected that is now
associated with a user in order to register device callbacks.

Fixes: 146521191
Test: Associate without error and unit tests pass
Change-Id: Id7514034c5508d6a73be289bf3dbb28baa8285d7