Remove Android.mk versions of modules that have been converted to Android.bp
The Android.mk versions of modules that have been converted to
Android.bp shouldn't be necessary, remove them.  The only reason
to keep both seems to be b/113294940, which has an easy workaround.

Also convert car-media-common and its dependency car-arch-common,
which would otherwise need the workaround, as well as
CarMediaCommonRoboTests.

Test: m checkbuild
Test: m RunCarArchCommonRoboTests RunCarAppsCommonRoboTests
Test: m RunCarMediaCommonRoboTests
Change-Id: Id6491d2c254f691d2a410a59d50352f60b7ff352