Automatically provide Toolbar in Activities
First, break up toolbar into ToolbarController and Toolbar

We want to be able to use the Toolbar with any layout, not
just one that is inflated in a custom view. So we added
ToolbarController, a new class that doesn't extend View,
to act as the new Toolbar interface. The old Toolbar
class still exists, but now just delegates to ToolbarController.

Then add a CarUiAppCompatActivity that creates a "sub-decor"
view with the toolbar, and use a ToolbarController with
the sub-decor.

Test: Manually
Change-Id: I2be1dfca5afd5379b1e7237cd3c8cb33dc00bc21