Remove BluetoothMapClient refs from car-messenger-common
CarMessenger never migrated to use the shared libraries and utils in
car-messenger-common, it will in the future on Master after Q has closed.
For now, we can remove Car Messenger specific APIs from this lib in order
to make backporting to Q much easier. (Check bug for more details)

Change-Id: I61166f9a0d39bb755dc042fa70e9057717a5672b
Fix: 148160399
Test: mma CDS app