Exclude dismissed and read messages from the notification
When a user dismisses a message notification, OR when a message
notification is successfully read aloud, the messages in that notification
should not appear when/if that notification gets updated with newer
messages.

This change combines the two states (dismiss and read) since they both
have the same logical consequence.

Bug: 146362547
Test: manual
Change-Id: I4841ba5641fdb8682639ecec07f2f6a510076a1a