Add insets to base layouts
Because features of the base layout can partially cover
the content of the application, the application needs to know how
to indent it's content that the user interacts with, but not
the content that is in the background.

This can now be done by overriding onCarUiInsetsChanged(Insets) from
CarUiAppCompatActivity. The default implmentation of that method
will apply insets as padding to the content view, making all content
in an app fit within the insets.

The OEM can specify the insets of their base layout by adding the tags
"car_ui_(left/right/bottom/top)_inset" to views. The insets will be
set to values such that content won't be obscured by the tagged views.
Tags are used because they can be applied to views that already have
an id used for something else. Like in our reference base layout,
a view has both the id car_ui_toolbar_background and the tag
car_ui_top_inset.

Fixes: 149421127
Test: Manually
Change-Id: Ic0fdcd71abd6e717691c7a3828be2d4d17659365