DO NOT MERGE: Redesign base layouts
Previously we had a CarUiAppCompatActivity that all other
activities had to subclass from in order to get the base layout.
Since several scenarios where that is not feasible were found,
switch to a new approach.

The new approach registers for activity lifecycle callbacks,
so CarUi receives a reference to every activity during its
onCreate(). The base layout is installed during that callback.

To register for lifecycle callbacks, we created a content
provider. Content providers are initialized very early on
in the application, before the first activity is launched.
An app only needs to include the content provider to set up
base layouts. Currently we require apps to add the <provider>
tag to their AndroidManifest.xml to enable our content provider
and use base layouts. However, when we are confident that this
doesn't significantly effect app startup time, we will put the
<provider> tag in our library's AndroidManifest.xml so it will
automatically be merged into the app's. This auto merging
only works properly on R though (b/113294940). On Q apps will
still have to add it manually.

This CL also allows preference fragments to use the base layout.
It keeps the old preference layout using the old style toolbar
and only uses the base layout version when using
Theme.CarUi.WithToolbar. This is for backwards-compatibility
with non-baselayout apps.

Also update all screens of paintbooth to use base layouts.

Fixes: 150237357
Test: Manually - checked media, dialer, and paintbooth
Change-Id: I0b92300b07c29cb43edc560f961d09c006df35cd