DO NOT MERGE Removing nested recycler view from CarUiRecycelrView.
This change will help not to delegate each and every call from the parent to the nested RV. Now we will dettach the inflated CarUiRV and attach a container with the scrollbar and the current RV with the same Layoutparams as before.

Also, car_ui_scrollbar_gutter and car_ui_scrollbar_position are noe removed as we can achieve the gutter by setting the start and end margin on the recyclerView within "car_ui_recycler_view.xml". Position of scrollbar can also be modified in the same file using RROs.

Bug: 150953626
Test: Robotest
Change-Id: I2617e4ab7a728d69b6b910129f3880972179ac33