Set channelId in Notification Builder every time.
If a projection app is active in foreground, all Notifications should
be posted silently. When the projection app is in background,
notifications should show with a beep and with a HUN. This is
controlled via Notification Channel IDs, where one is for the silent
notifications, and other is for the HUNs + beep.

This means every time a conversation is updated, the notification
channel id might have changed since the last notification post. This
change addresses this.

Bug: 138747359
Test: verified on hawk. CDS + AAP was not working together well, so I
mocked the changing of the Projection Status, verified that convos
posted with HUNs when Projection is not active, and the same + new
convos posted silently when Projection is in background. Also tested
the other way around.

Change-Id: I2f4236e6a2f85feb34d4ddd06b8c8e0484112934