Prevent clicking through the Toolbar in base layouts
The old style toolbar was a custom view that included overrides for
onTouchEvent() and onHoverEvent() to prevent clicking through it
to views underneath it. Now with base layouts, there is no custom
view that can do that.

Add a ClickBlockingView that can be added to base layouts. This view
prevents any touches from passing through it.

Fixes: 153904642
Test: Manually
Change-Id: I7d65ca7b6abdf5c15fce15ecb319c745b03a3d12