Create encryption-runner library within car/libs.
This library is a fork of the `EncryptionRunner` within
packages/services/Car. That library will be removed in master so
branching in rvc-dev so that development can upstream to master.

The library cannot be moved in rvc-dev because there are still
classes within packages/services/Car that depend on it.

Bug: 155198346
Test: m -j
Test: Run connected-device-lib-unit-tests
Change-Id: I3530034e476888954ad4c5a97af2f965ed273d07