DO NOT MERGE Fix paddings for RTL
Using paddingStart/End along with padding can result inconsistency between the layouts of apps with or without RTL support.

When RTL is disabled, padding/paddingHorizontal overrides the value of paddingStart/End.
When RTL is enabled, paddingStart/End overrides the value of padding/paddingHorizontal.

This CL updates the styles used in the Car/libs to make sure we get the
same padding for both cases.

Bug: 155687655
Test: Manual on Hawk/Simulator
Change-Id: Iffb56c00c74990f6ca0342b7faac47a50d555b89