Fix failing toolbar robolectric tests
We changed certain views to be inside a container, and
show/hide the container. So we can't just simply call
getVisibility() on the views. We should use .isShown(),
but that returns false if the views aren't attached to
a real view hierarchy. Ours aren't because we make a
new toolbar with a spied context. So this cl creates
an alternate version of isShown that just checks up
to the toolbar.

Fixes: 159340746
Test: atest CarUiRoboTests
Change-Id: Ic503b187e6d762fd8589ece3829047a781710bc5