DO NOT MERGE: Update base layouts to match R
CompanionDeviceSupport is backporting their whole
app from R -> Q as part of b/159485985. This means
they're now our only user of BaseLayouts on Q.

We need to update baselayouts to include some of
the fixes we made on R, most notably switching
back to the legacy toolbar on Q automatically,
so that OEM customizations of the legacy toolbar
are applied to apps that use the base layout
toolbar as well.

Fixes: 160278133
Test: Manually, paintbooth
Change-Id: I6687addadb6a7cc33190ed226bf87d9d2a4b816d