Fix baselayout inset calculation
A change was made recently to support Media's unconventional
base layout implementation, and it didn't work correctly.
It was causing extra paddings to be applied to many apps
under certain conditions.

The issue revolves around how in most apps, the content view
container and the content view have the same size and position.
But in MediaCenter that's not necessarily the case.

Update the inset calculation code to take the difference
between the content view and the content view container
into account when they are different.

Fixes: 160901371
Test: Manually, in paintbooth and media, and with
      the VerticalToolbarRRO on both of them
Change-Id: I364c488fa8b2f618d4e42e76cd5ae70bc886dcdc