DO NOT MERGE Update the AndroidManifest template to include bypass IdMap gen
Setting the category to "BypassIdMapV1" will cause these static RROs not
to have an IDMAP V1 generated within the PackageManager which improves
boot performance. IDMAP v1 allowed for overlays of values within an
AndroidManifest file but that functionality and IDMAP v1 has been
removed in the next version of Android and is not used here.

Bug: 159478185
Test: Manualcar-ui-lib/referencedesign/AndroidManifest.xml
Change-Id: I02dc920fd8c7d4da2ae487e7001273903de730f5