Make rotary scrolling opt-in for RecyclerViews
Rotary scrolling isn't needed for most CarUiRecyclerViews because all
the items are focusable. It causes problems when there's a gap between
adjacent rows, as is the case in the launcher. This CL makes rotary
scrolling opt-in rather than opt-out for CarUiRecyclerViews. We can
enable rotary scrolling where we need it, either declaratively or
programmatically via a new method in CarUiUtils.

Test: try scrolling in Launcher and Settings
Test: try scrolling in scroll tab of RotaryPlayground
Test: set contentDescription programmatically to enable rotary scrolling
Test: make RotaryPlayground enable rotary scrolling programmatically
Bug: 160922045
Change-Id: Ia6ed5ddafedd9c628cb9d5430fd92b1388215ea8