DO NOT MERGE: Add setAllowDismissButton to AlertDialogBuilder
By default, the AlertDialogBuilder may add a "Dismiss" button if a
positive/negative/neutral button is not provided. This is so that
the dialog is still dismissible using the rotary controller. If
however, there are buttons that can close the dialog via
setAdapter(CarUiListItemAdapter) or a similar method, then it may
be desirable to suppress the addition of the dismiss button, which
this method allows for.

Test: Manually
Bug: 171067302
Change-Id: If84b6c2dee449931852400d594a555d835b3e23f