Fix recyclerview jank when using baselayouts
The base layout insets were dispatching after
the screen was already shown, so it was possible
to see the recyclerview behind the toolbar for a
second before it jumped to the correct position.

Don't wait for a global layout to recalculate the
insets, and scroll to the top in setPadding()
instead of waiting for the first global layout.

Fixes: 172575969
Test: Manually and atest CarUILibUnitTests
Change-Id: Ic5fa0b42229a58bbd2934a121d1bbf6762855186