Improve performance of toolbar.
TextView#setTitle does a lot of heavy lifting - there is measuring, drawing and object allocations, all of which run on the main thread. Hence, this method results in slowing the UI thread a bit. Use PrecomputedText that measure and calculates on background thread saving us time.

ImageView#setImageDrawable is bit slow because it does decoding on UI thread, We can solve this problem using worker thread or AsyncTaks.

Bug: 172746780
Test: Manual
Change-Id: Id51b44e6f21127a4f019a0eb56d386378d6849d0