Use AsyncTaskLoader to load database
After cursor is loaded, it also converts the cursor to the desired type
of data and then closes the cursor. The existing approach which uses the
AysncQueryHandler posts the curosr to main thread and then we start the
data convertion async which may cause the cursor not to close properly.

Test: manually for all pages
Bug: 169014584
Change-Id: I1b6669086efda0ee45d863c75fed05f93875c021