Remove text button container from MenuItems
Right now, with bordered MenuItems, the rotary
highligh is a box around the bordered button
in a way that doesn't quite line up with it.
To fix that, we should be focusing the actual
clickable buttons, not containers, and then
setting the rotary highlight correctly on
those buttons.

This cl removes the text container, and makes
the focus apply directly to each button/switch.
The icon is a little different because it
has both a foreground and background ImageView.

Note that this cl reverts the blue rotary
box around the textbutton to a simple ripple.
We should update the buttonStyle and switchStyle
in Theme.DeviceDefault to include a rotary
highlight that makes sense for them.

Test: Manually
Bug: 174490425
Change-Id: Ic047ddbb57263ec430dcf55cdb03ff158200f5b6