Improve IME opening logic
Currently, the IME opens when the search box is focused and it closes
when the search box loses focus. This prevents the user from navigating
to the Clear button because focus moves to the IME when it opens. This
change makes the IME open when the search box is clicked. It also moves
focus to the search box and opens the IME when the Clear button is
pressed. The logic to close the IME remains unchanged.

Test: Search for a contact using touch and rotary
Bug: 174768762
Change-Id: I09f134e3754adeb3b5bff5a12d5e0aac3661bf20