Refactor browse UI in preparation for wide screen support
This large change is necessary to support widescreen keyboard and
also to react to parts of the hierarchy becoming obsolete (bluetooth).

Previously, the MediaBrowserViewModel was assuming there would only
ever be a single list of media items the UI was interested in.

This change replaces the MediaBrowserViewModel by a new
MediaItemsRepository that allows to track changes in several lists
of media items (including the search results).


Fixes: 177097898 Remove mConnectedMediaBrowser
Fixes: 145758997 clean up MediaBrowserViewModel instantiation

Test: manual + make -j64 RunCarMediaCommonRoboTests

Change-Id: Iba1c6591029b55af8deb14b567b80fdad93db071
Merged-In: Iba1c6591029b55af8deb14b567b80fdad93db071