Unbundle car-ui-lib's soong build
While we don't have any dependencies
on unbundled APIs, car-ui-lib was still
compiling against the platform APIs. This
cl changes it to compile against only the
publicly released SDKs.

This means that we aren't able to access
@hide code anymore, which is good because
our google3 and gradle builds already
weren't able to. Recently a change was
submitted mistakingly adding usage of
an @hide api. This cl will help catch
that kind of mistake in the future.

Fixes: 175816471
Test: Manually
Change-Id: Iff158e34976cc8221a2a863eaacc5de47a8df45a