Remove usage of legacy toolbar from PreferenceFragment
If an app was using Theme.CarUi.NoToolbar, then the PreferenceFragment
would have a legacy toolbar included in it via the preference theme.
This is not desirable in all apps, so remove the preference layout-based
toolbar.

For apps to have a functioning preference fragment, they now need
to either use Theme.CarUi.WithToolbar, or add their own toolbar
and override getPreferenceToolbar().

Bug: 181694712
Test: Manually
Change-Id: I866bb0f19c216db8ed72589e402a38e5d8cbbb51