Deprecate toolbar state
Apps are beginning to request the ability to
show more combinations of toolbar elements,
like a logo while searching. To enable this,
make all the elements of a toolbar independently
settable, and when set the toolbar should do
its best to display them all.

SetState() always conflated multiple toolbar
elements together and decided what was and
wasn't shown. This imposed artificial
restrictions on what could be shown.

Deprecate setState and add setSearchMode()
and a NavButtonMode.DISABLED to replace it.

Test: Manually
Bug: 179520668
Change-Id: I8ca0771ef2a6d3c72b9281cb7d614aab8416f00f