Fix instrumentation tests
The new sparkle animation on ripples breaks espresso's
`onView(...).perform(click())`, as it presses down,
waits for idle, and then releases the press. The
sparkle animation causes the system to never go idle
after the down press happens, so espresso never
releases the button.

Adding --no-window-animation to the test runner
arguments will disable the sparkle animation, and
allow the click to work again.

Fixes: 184746725
Test: atest CarUiLibUnitTests and presubmits
Change-Id: I214dca8cb8867f8f9c9db81e77868d1b76008c7e