Add non-deprecated toolbar functions
Many of the methods in ToolbarController took arguments
that were inner classes of the deprecated Toolbar class.
This caused deprecation warnings whenever these methods
were called, even though they weren't directly deprecated.
Add alternatives to all of these methods that don't use
any deprecated arguments.

In addition to this, this cl also includes a new method
of specifying tabs. The old method was based on views,
which could theoretically allow the app to change the
tab styling in unintended ways.

The full list of methods changed:
- addTab, clearAllTabs, getTab, getTabPosition, getTabCount,
  registerOnTabSelectedListener, and
  unregisterOnTabSelectedListener were all replaced with
  setTabs()
- setNavButtonMode, getNavButtonMode -> setBackButtonMode
- registerOnSearchCompletedListener -> registerSearchCompletedListener
- unregisterOnSearchCompletedListener -> unregisterSearchCompletedListener
- registerOnBackListener -> registerBackListener
- unregisterOnBackListener -> unregisterBackListener

Fixes: 179195852
Test: atest CarUiLibUnitTests
Change-Id: Id631b7aadeae1e4709394bc38bfab61feefc957d