Fix scrolling issue when CarUiRV is scrolled and attached back to parent
CarUiRv always scrolls to the top i.e at 0th item when paddings are applied and the data has been scrolled yet. This is to inflate the data properly when CarUiRv is first inflated in the activity. But if we detach the view, scroll a bit and come back to data 0 and attach it back to the applications parent again then the data was off. We now check if the list is at the start and then scroll the data.

Bug: 186175785
Test: Manual
Change-Id: Ib4fd1793de5293d0b78b2a62eb3816e5c9aa7529