Highlight thumb when CarUiRecyclerView is focused
If rotary scrolling is enabled on a CarUiRecyclerView, the view itself
is focused when no focusable descendant views are visible. Before this
change, there was no visual indication of focus in this situation.
After this change, the thumb of the scrollbar is blue in this
situation, rather than the usual gray.

(Manual cherry-pick from ag/14832577.)

Test: manual - see bug for video
Bug: 178515570
Change-Id: I4cf6b5b1ec3115e87587cbddd005b324b61e1692