Adds collator field to Contact.java in car-telephony-common.
The current compareTo implementation creates for each comparision a new
Collator instance which will immediately gets deprecated again and needs
to be cleaned up by the Garbage Collector. If the Garbage Collector
needs to clean up all the instance the system might freeze. In this case
it will happen when >5000 Contacts needs to be sorted as around 75000
Collator instances are created.

This is an intermediate fix which might introduce other issue when the
collator should be changed, but the previous instance is still used.

Additionally fixing existing tests of ContactTest.java.

Bug: 187440159
Test: unit testing
Change-Id: I476133ac5b0f2713cb0f5cea6db25e662fda2942