Adds collator field to Contact.java in car-telephony-common
The current compareTo implementation creates for each comparision a new
Collator instance which will immediately gets deprecated again and needs
to be cleaned up by the Garbage Collector. If the Garbage Collector
needs to clean up all the instance the system might freeze. In this case
it will happen when >5000 Contacts needs to be sorted as around 75000
Collator instances are created.

Additionally fixing existing tests of ContactTest.java.

Bug: 190966644
Test: unit testing

Change-Id: Ice3936131b3faa1f8fb0362a4f2f50e0a9711b8c