Fix back button being visible but not clickable when not using setState
The nav button started visible due to the way the layout file
was set up, and required update() to be called to hide it again.
setState() would call update() the first time it's called,
but setNavButtonMode(DISABLED) would not, because the toolbar
starts with a disabled NavButtonMode and assumes there's
no change.

Remove the back button from the default resources.

Bug: 179177137
Test: Manually
Change-Id: Id9301754a326240884f6fbc5c1d68ee2d770593b