Add setSecure() to list items
PermissionController needs to not allow the grant permission
buttons to be clicked when there is a full or partial overlay
onscreen, as it could trick users into granting a permission
they didn't mean to. The only way we could find to detect
an overlay was via a View api, so chassis has to handle it.

Add a setSecure() method that when set, will cause the
CarUiListItem to not respond to touches when there is an
overlay onscreen.

Bug: 192088266
Test: atest CarUILibUnitTests
Change-Id: Ib30d2c7599e14c6062e912345eed2bb69387e19b