Add ability to control tab tint and hide plugin toolbar background
These are both for MediaCenter, which uses these features currently
via the deprecated tabs apis and the static toolbar.

Media center uses the ability to control the icon tint to tint
BitmapDrawables red, as a debugging tool.

Media center hides the toolbar background on the now playing
page, to make it blend in better with the rest of the screen.

Bug: 192289281
Bug: 187078438
Test: Manually
Change-Id: I2de7ab9a5bd3a7b3c322b569af00935d608feecc