Initialize focus inside the LazyLayoutView properly
When initializing the focus, if the LazyLayoutView is not visible,
wait until it's visible or after a timeout. This is useful when
view animation is involved.

Change findRotaryContainer() to return the first loaded container.
This is to fix a bug in Media, where there are two visible rotary
containers in browsing mode: one is empty and in the root
controller, another one is loaded, in the browse controller, and
on top of the former one. In this case it should return the later
one.

Remove willNotDraw() check because it shouldn't be used to decide
whether a view is visible.

Fixes: 197590804
Fixes: 198509791
Fixes: 198487444
Test: manual

Change-Id: Idfa92efbd0f020d720373c4789fd6576cacece93